Added an avr switch mechanism
