+++ avrs/arduboy/avr_config.ml
@@ -1,0 +1,4 @@
+let default_mmcu  = "atmega32u4"
+let default_avr   = "avr109"
+let default_baud  = 57_600
+let default_clock = 16_000_000
