+++ avrs/arduino_uno/avr_config.ml
@@ -1,0 +1,4 @@
+let default_mmcu  = "atmega328p"
+let default_avr   = "arduino"
+let default_baud  = 115_200
+let default_clock = 16_000_000
