+++ src/omicrob/avr_config.ml
@@ -1,0 +1,1 @@
+../../avrs/arduboy/avr_config.ml
